# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Food Delivery API
  version: 1.0.0
  description: REST API for a food delivery service.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /api/v1/auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      description: Creates a new user account with the user role.
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              username: john
              email: john@example.com
              password: password123
      responses:
        "201":
          description: User successfully registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request body or validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: user already exists
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: internal server error
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      summary: Log in
      description: Authenticates a user and returns a JWT access token.
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: john@example.com
              password: password123
      responses:
        "200":
          description: Authentication successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid request body or validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid email or password.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: invalid credentials
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: internal server error
...
  /api/v1/addresses:
    post:
      tags:
        - Addresses
      summary: Create an address
      description: Creates a saved delivery address for the authenticated user.
      operationId: createAddress
      security:
        - BearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAddressRequest"
            example:
              label: Home
              city: Amsterdam
              street: Damrak
              house_number: "10"
              apartment_number: "42"
              entrance: "2"
              floor: "5"
              comment: Call before delivery

      responses:
        "201":
          description: Address successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"

        "400":
          description: Invalid request body or address data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: unauthorized

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags:
        - Addresses
      summary: List user addresses
      description: Returns saved addresses belonging to the authenticated user.
      operationId: listUserAddresses
      security:
        - BearerAuth: []

      responses:
        "200":
          description: Address list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"

        "400":
          description: Invalid address data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/addresses/{id}:
    parameters:
      - $ref: "#/components/parameters/AddressId"

    get:
      tags:
        - Addresses
      summary: Get address by ID
      description: Returns an address belonging to the authenticated user.
      operationId: getAddressById
      security:
        - BearerAuth: []

      responses:
        "200":
          description: Address found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"

        "400":
          description: Invalid address ID or address data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "404":
          description: Address not found or does not belong to the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: address not found

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      tags:
        - Addresses
      summary: Update an address
      description: Partially updates an address belonging to the authenticated user.
      operationId: updateAddress
      security:
        - BearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAddressRequest"
            example:
              label: Office
              comment: Leave at reception

      responses:
        "200":
          description: Address successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"

        "400":
          description: Invalid address ID, request body, or address data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "404":
          description: Address not found or does not belong to the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: address not found

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Addresses
      summary: Delete an address
      description: Deletes an address belonging to the authenticated user.
      operationId: deleteAddress
      security:
        - BearerAuth: []

      responses:
        "204":
          description: Address successfully deleted.

        "400":
          description: Invalid address ID or address data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "404":
          description: Address not found or does not belong to the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: address not found

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    AddressId:
      name: id
      in: path
      required: true
      description: Address identifier.
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          description: User name. Leading and trailing whitespace is removed.
          example: john
        email:
          type: string
          description: User email. Leading and trailing whitespace is removed.
          example: john@example.com
        password:
          type: string
          minLength: 8
          writeOnly: true
          description: Password must contain at least 8 characters.
          example: password123
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          description: User email.
          example: john@example.com
        password:
          type: string
          writeOnly: true
          example: password123
    UserRole:
      type: string
      enum:
        - user
        - admin
      example: user
    User:
      type: object
      required:
        - id
        - username
        - email
        - role
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: john
        email:
          type: string
          example: john@example.com
        role:
          $ref: "#/components/schemas/UserRole"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LoginResponse:
      type: object
      required:
        - access_token
        - user
      properties:
        access_token:
          type: string
          description: JWT access token.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: "#/components/schemas/User"
...
    CreateAddressRequest:
      type: object
      required:
        - city
        - street
        - house_number
      properties:
        label:
          type:
            - string
            - "null"
          example: Home
        city:
          type: string
          minLength: 1
          example: Amsterdam
        street:
          type: string
          minLength: 1
          example: Damrak
        house_number:
          type: string
          minLength: 1
          example: "10"
        apartment_number:
          type:
            - string
            - "null"
          example: "42"
        entrance:
          type:
            - string
            - "null"
          example: "2"
        floor:
          type:
            - string
            - "null"
          example: "5"
        comment:
          type:
            - string
            - "null"
          example: Call before delivery
    UpdateAddressRequest:
      type: object
      properties:
        label:
          type:
            - string
            - "null"
        city:
          type:
            - string
            - "null"
          minLength: 1
        street:
          type:
            - string
            - "null"
          minLength: 1
        house_number:
          type:
            - string
            - "null"
          minLength: 1
        apartment_number:
          type:
            - string
            - "null"
        entrance:
          type:
            - string
            - "null"
        floor:
          type:
            - string
            - "null"
        comment:
          type:
            - string
            - "null"
    Address:
      type: object
      required:
        - id
        - user_id
        - city
        - street
        - house_number
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          example: 1
        user_id:
          type: integer
          format: int64
          example: 5
        label:
          type:
            - string
            - "null"
          example: Home
        city:
          type: string
          example: Amsterdam
        street:
          type: string
          example: Damrak
        house_number:
          type: string
          example: "10"
        apartment_number:
          type:
            - string
            - "null"
          example: "42"
        entrance:
          type:
            - string
            - "null"
        floor:
          type:
            - string
            - "null"
        comment:
          type:
            - string
            - "null"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...